Remove casts now that sizes use the "long" type.
These are no longer necessary, and they sort of defeat the purpose of
using the "long" type for some of the sizes if some of the size values
are going to be cast to the "int" type throughout the execution.

+++ src/org/isep/simizer/example/consistency/OptimisticPolicy.java
@@ -59,7 +59,7 @@ public class OptimisticPolicy extends StoreApplication {
     } else if (res.getVersion() < val) {
       res.setVersion(val);
     }
-    scheduler.write(res, (int) res.size());
+    scheduler.write(res, res.size());
 
     // asynchronous, we reply before replication
     sendResponse(scheduler, request);
@@ -93,7 +93,7 @@ public class OptimisticPolicy extends StoreApplication {
     
     // /!\ must be atomic
     if (localResource != null && localResource.getVersion() < requestResource.getVersion()) {
-      scheduler.write(requestResource, (int) requestResource.size());
+      scheduler.write(requestResource, requestResource.size());
     }
 
     return request;
